refactor(route_extractor): move the RouteExtrator class into de Extractors module.
